refactor: Extract anchor debug plugin.
